+++ src/syslog2irc/router.py
@@ -13,6 +13,8 @@ from dataclasses import dataclass
 import logging
 from typing import Any, Dict, Optional, Set
 
+from .network import format_port, Port
+
 
 logger = logging.getLogger(__name__)
 
@@ -21,14 +23,14 @@ logger = logging.getLogger(__name__)
 class Route:
     """A route from a syslog message receiver port to an IRC channel."""
 
-    syslog_port: int
+    syslog_port: Port
     irc_channel_name: str
 
 
 class Router:
     """Map syslog port numbers to IRC channel names."""
 
-    def __init__(self, ports_to_channel_names: Dict[int, Set[str]]) -> None:
+    def __init__(self, ports_to_channel_names: Dict[Port, Set[str]]) -> None:
         self.ports_to_channel_names = ports_to_channel_names
         self.channel_names_to_ports = map_channel_names_to_ports(
             ports_to_channel_names
@@ -51,17 +53,17 @@ class Router:
         logger.info(
             'Enabled forwarding to IRC channel %s from syslog port(s) %s.',
             channel_name,
-            ', '.join(map(str, sorted(ports))),
+            ', '.join(map(format_port, sorted(ports, key=lambda p: p.number))),
         )
 
     def is_channel_enabled(self, channel: str) -> bool:
         return channel in self.enabled_channels
 
-    def get_channel_names_for_port(self, port: int) -> Set[str]:
+    def get_channel_names_for_port(self, port: Port) -> Set[str]:
         return self.ports_to_channel_names[port]
 
 
-def map_ports_to_channel_names(routes: Set[Route]) -> Dict[int, Set[str]]:
+def map_ports_to_channel_names(routes: Set[Route]) -> Dict[Port, Set[str]]:
     ports_to_channel_names = defaultdict(set)
     for route in routes:
         ports_to_channel_names[route.syslog_port].add(route.irc_channel_name)
@@ -69,8 +71,8 @@ def map_ports_to_channel_names(routes: Set[Route]) -> Dict[int, Set[str]]:
 
 
 def map_channel_names_to_ports(
-    ports_to_channel_names: Dict[int, Set[str]]
-) -> Dict[str, Set[int]]:
+    ports_to_channel_names: Dict[Port, Set[str]]
+) -> Dict[str, Set[Port]]:
     channel_names_to_ports = defaultdict(set)
     for port, channel_names in ports_to_channel_names.items():
         for channel_name in channel_names: